Add player movement limits and game manager comments
Added player movement limits (currently uses magic values rather than retrieving pre-defined values).
Added some documentation comments to the GameManagerScript and also added some empty methods to plan for the future.
Added a LeaderboardScore class for when the leaderboard will be implemented for what needs to be in the class currently.